optimize: new AttachmentDirectorySelectDialog.vue and optimize Attachments.vue
