+++ console/src/utils/api-client.ts
@@ -6,8 +6,6 @@ import {
 	SettingIkarosRunV1alpha1ConfigmapApi,
 	V1alpha1EpisodeCollectionApi,
 	V1alpha1SubjectCollectionApi,
-	V1alpha1FileApi,
-	V1alpha1FolderApi,
 	V1alpha1IndicesApi,
 	V1alpha1PluginApi,
 	V1alpha1SubjectApi,
@@ -161,8 +159,6 @@ function setupApiClient(axios: AxiosInstance) {
 			baseURL,
 			axios
 		),
-		file: new V1alpha1FileApi(undefined, baseURL, axios),
-		folder: new V1alpha1FolderApi(undefined, baseURL, axios),
 		subject: new V1alpha1SubjectApi(undefined, baseURL, axios),
 		subjectSyncPlatform: new V1alpha1SubjectSyncPlatformApi(
 			undefined,
